+++ localflavor/ie/forms.py
@@ -37,12 +37,12 @@ class EircodeField(RegexField):
         # https://www.eircode.ie/docs/default-source/Common/prepareyourbusinessforeircode-edition3published.pdf?sfvrsn=2
         value = super().to_python(value)
         if value in self.empty_values:
-            return self.empty_value
+            return value
         return value.upper().replace(" ", "")
 
     def prepare_value(self, value):
         # Display the Eircode with a space.
         value = self.to_python(value)
         if value in self.empty_values:
-            return self.empty_value
+            return value
         return '{} {}'.format(value[0:3], value[3:8])