+++ localflavor/mx/forms.py
@@ -112,7 +112,7 @@ class MXRFCField(RegexField):
     def clean(self, value):
         value = super().clean(value)
         if value in self.empty_values:
-            return self.empty_value
+            return value
         value = value.upper()
         if self._has_homoclave(value):
             if not value[-1] == self._checksum(value[:-1]):
@@ -194,7 +194,7 @@ class MXCLABEField(RegexField):
     def clean(self, value):
         value = super().clean(value)
         if value in self.empty_values:
-            return self.empty_value
+            return value
         if not value.isdigit():
             raise ValidationError(self.error_messages['invalid'], code='invalid')
         if not self._checksum(value):
@@ -245,7 +245,7 @@ class MXCURPField(RegexField):
     def clean(self, value):
         value = super().clean(value)
         if value in self.empty_values:
-            return self.empty_value
+            return value
         value = value.upper()
         if value[-1] != self._checksum(value[:-1]):
             raise ValidationError(self.error_messages['invalid_checksum'], code='invalid_checksum')
@@ -303,7 +303,7 @@ class MXSocialSecurityNumberField(RegexField):
     def clean(self, value):
         value = super().clean(value)
         if value in self.empty_values:
-            return self.empty_value
+            return value
         if value[-1] != self.__checksum(value[:-1]):
             raise ValidationError(self.error_messages['invalid_checksum'], code='invalid_checksum')
         return value